#62: Refactored header and footer for nanomine
