+++ app/src/components/nanomine/PageFooter.vue
