+++ app/src/components/nanomine/PageHeader.vue
